test-setup.sh: Allow re-using existing container

+++ test-setup.sh
@@ -45,31 +45,42 @@ name_matches=$(docker ps -a --filter name="^$TESTDB_CONTAINER$" | wc -l)
 if [[ "$name_matches" -gt 1 ]]; then
   # We always get one header line. If there are more lines, then the container exists.
   echo "Container $TESTDB_CONTAINER exists and will be destroyed."
-  read -rp "Continue? [y/N]: " response
-  if [[ $response != "y" && $response != "Y" ]]; then
+  echo "Or answer [r]euse to keep the existing one."
+
+  read -rp "Continue? [y/N/r]: " response
+  if [[ $response == y || $response == Y ]]; then
+    reuse_container=false
+  elif [[ $response == r || $response == R ]]; then
+    reuse_container=true
+  else
     echo "Aborting script"
     exit 2
   fi
-  # Destroy it.
-  echo -n "Destroying $TESTDB_CONTAINER... "
-  docker rm --force "$TESTDB_CONTAINER" >/dev/null
-  echo "done"
+
+  if [[ $reuse_container != true ]]; then
+    # Destroy it.
+    echo -n "Destroying $TESTDB_CONTAINER... "
+    docker rm --force "$TESTDB_CONTAINER" >/dev/null
+    echo "done"
+  fi
 fi
 
-# Build the image.
-echo -n "Rebuilding DB image... "
-docker build -t "$TESTDB_IMAGE" ./db >/dev/null
-echo "done"
+if [[ $reuse_container != true ]]; then
+  # Build the image.
+  echo -n "Rebuilding DB image... "
+  docker build -t "$TESTDB_IMAGE" ./db >/dev/null
+  echo "done"
 
-# Run the container.
-echo -n "Launching container... "
-docker run --rm -d \
-  --env "MONGODB_REPLICA_SET_MODE=primary" \
-  --env "MONGODB_REPLICA_SET_KEY=$TESTDB_PASSWORD" \
-  --env "MONGODB_ROOT_PASSWORD=$TESTDB_PASSWORD" \
-  --name "$TESTDB_CONTAINER" \
-  "$TESTDB_IMAGE" >/dev/null
-echo "done"
+  # Run the container.
+  echo -n "Launching container... "
+  docker run --rm -d \
+    --env "MONGODB_REPLICA_SET_MODE=primary" \
+    --env "MONGODB_REPLICA_SET_KEY=$TESTDB_PASSWORD" \
+    --env "MONGODB_ROOT_PASSWORD=$TESTDB_PASSWORD" \
+    --name "$TESTDB_CONTAINER" \
+    "$TESTDB_IMAGE" >/dev/null
+  echo "done"
+fi
 
 # Get the IP.
 echo -n "Getting IP address... "
@@ -79,32 +90,34 @@ mongo_uri="mongodb://root:$TESTDB_PASSWORD@$ip_addr"
 echo "done"
 echo "Container $TESTDB_CONTAINER running at $ip_addr"
 
-# Configure the replica set hostname.
-# This defaults to the container hostname (container ID), but must be routable
-# from this host, so set it to the container IP.
-# This only works once mongodb is fully up and ready, which is difficult to
-# detect exactly, so wait for the port to be available and then retry in a loop.
-echo -n "Configuring replica set... "
-./db/wait-for-it.sh "$ip_addr:27017" &>/dev/null
-remaining_attempts=10
-set +e
-trap - ERR
-while true; do
-  ((remaining_attempts--))
-  if [[ $remaining_attempts -eq 0 ]]; then
-    # Die if the last attempt fails.
-    set -e
-    trap "$errtrap" ERR
-  fi
-  mongosh "$mongo_uri" \
-    --eval "cfg = rs.conf(); cfg.members[0].host = '$ip_addr'; rs.reconfig(cfg)" \
-    &>/dev/null
-  # An `&& break` would stop the mongosh invocation from triggering `set -e`, so
-  # test the return code afterwards.
-  test $? -eq 0 && break
-  sleep 1
-done
-echo "done"
+if [[ $reuse_container != true ]]; then
+  # Configure the replica set hostname.
+  # This defaults to the container hostname (container ID), but must be routable
+  # from this host, so set it to the container IP.
+  # This only works once mongodb is fully up and ready, which is difficult to
+  # detect exactly, so wait for the port to be available and then retry in a loop.
+  echo -n "Configuring replica set... "
+  ./db/wait-for-it.sh "$ip_addr:27017" &>/dev/null
+  remaining_attempts=10
+  set +e
+  trap - ERR
+  while true; do
+    ((remaining_attempts--))
+    if [[ $remaining_attempts -eq 0 ]]; then
+      # Die if the last attempt fails.
+      set -e
+      trap "$errtrap" ERR
+    fi
+    mongosh "$mongo_uri" \
+      --eval "cfg = rs.conf(); cfg.members[0].host = '$ip_addr'; rs.reconfig(cfg)" \
+      &>/dev/null
+    # An `&& break` would stop the mongosh invocation from triggering `set -e`, so
+    # test the return code afterwards.
+    test $? -eq 0 && break
+    sleep 1
+  done
+  echo "done"
+fi
 
 echo
 echo "Please set 'export ROCKET_DB_URI=$mongo_uri' in your shell"
